Moving logging functions into phdi_google_cloud_utils
